+++ components/mainPage/Bio.jsx
