+++ components/mainPage/Footer.jsx
