+++ components/mainPage/Intro.jsx
