+++ components/mainPage/NavBar.jsx
